split partition generate logical
