+++ src/service/search/partition/mod.rs
@@ -15,16 +15,18 @@
 
 pub(crate) mod aggregate;
 pub(crate) mod cpu_cores;
+pub(crate) mod histogram;
+pub(crate) mod regular_partition;
 pub(crate) mod settings;
 pub(crate) mod sql_context;
 pub(crate) mod stream_files;
 
-use std::cmp::max;
-
-use config::{get_config, meta::sql::OrderBy};
+use config::get_config;
 
 use crate::service::search::partition::{
-    settings::PartitionSettings, sql_context::PartitionSqlContext,
+    histogram::generate_partitions_aligned_with_histogram_interval,
+    regular_partition::generate_partitions_with_mini_partition, settings::PartitionSettings,
+    sql_context::PartitionSqlContext,
 };
 
 /// Generate partitions for a time range derived from the SQL context.
@@ -55,840 +57,5 @@ pub fn generate_partitions(
             order_by,
             mini_partition_duration_secs,
         )
-    }
-}
-
-/// Generate partitions with a mini partition for faster initial results
-fn generate_partitions_with_mini_partition(
-    start_time: i64,
-    end_time: i64,
-    step: i64,
-    order_by: OrderBy,
-    mini_partition_duration_secs: u64,
-) -> Vec<[i64; 2]> {
-    let mut partitions = Vec::new();
-    let mini_partition_size = calculate_mini_partition_size(step, mini_partition_duration_secs);
-
-    log::debug!(
-        "mini_partition_size_microseconds: {mini_partition_size}, step: {step}, end_time: {end_time}, start_time: {start_time}"
-    );
-
-    // Add mini partition if possible
-    if let Some(mini_partition) =
-        create_mini_partition(start_time, end_time, mini_partition_size, order_by)
-    {
-        partitions.push(mini_partition);
-
-        // Generate remaining partitions
-        match order_by {
-            OrderBy::Desc => {
-                let remaining_start = mini_partition[0];
-                let remaining_partitions =
-                    generate_regular_partitions(start_time, remaining_start, step, OrderBy::Desc);
-                partitions.extend(remaining_partitions);
-            }
-            OrderBy::Asc => {
-                let remaining_end = mini_partition[1];
-                let remaining_partitions =
-                    generate_regular_partitions(remaining_end, end_time, step, OrderBy::Asc);
-                partitions.extend(remaining_partitions);
-            }
-        }
-    } else {
-        // Fall back to regular partitions if mini partition can't be created
-        partitions = generate_regular_partitions(start_time, end_time, step, order_by);
-    }
-
-    handle_empty_partitions(&mut partitions, start_time, end_time);
-    partitions
-}
-
-/// Generate partitions aligned with the histogram interval
-fn generate_partitions_aligned_with_histogram_interval(
-    start_time: i64,
-    end_time: i64,
-    step: i64,
-    order_by: OrderBy,
-    min_step: i64,
-) -> Vec<[i64; 2]> {
-    let mut partitions =
-        generate_histogram_aligned_partitions(start_time, end_time, step, order_by, min_step);
-
-    handle_empty_partitions(&mut partitions, start_time, end_time);
-    partitions
-}
-
-/// Generate histogram-aligned partitions
-fn generate_histogram_aligned_partitions(
-    start_time: i64,
-    end_time: i64,
-    step: i64,
-    order_by: OrderBy,
-    min_step: i64,
-) -> Vec<[i64; 2]> {
-    let mut partitions = Vec::new();
-    let duration = end_time - start_time;
-
-    // Handle the alignment for the first and last partition
-    let mut new_end = end_time;
-    let last_partition_step = end_time % min_step;
-    if last_partition_step > 0 && duration > min_step {
-        new_end = end_time - last_partition_step;
-        partitions.push([new_end, end_time]);
-    }
-    let mut new_start = start_time;
-    let last_partition_step = start_time % min_step;
-    if last_partition_step > 0 && duration > min_step {
-        new_start = start_time + min_step - last_partition_step;
-        partitions.push([start_time, new_start]);
-    }
-
-    let mut end = new_end;
-    while end > new_start {
-        let start = max(end - step, new_start);
-        partitions.push([start, end]);
-        end = start;
-    }
-
-    // We need to reverse partitions if query is ASC order
-    partitions.sort_by(|a, b| b[0].cmp(&a[0]));
-    if order_by == OrderBy::Asc {
-        partitions.reverse();
-    }
-
-    partitions
-}
-
-/// Handle edge case of empty partitions
-fn handle_empty_partitions(partitions: &mut Vec<[i64; 2]>, start_time: i64, end_time: i64) {
-    if partitions.is_empty() {
-        partitions.push([start_time, end_time]);
-    }
-}
-
-/// Calculate mini partition size in microseconds
-fn calculate_mini_partition_size(step: i64, mini_partition_duration_secs: u64) -> i64 {
-    std::cmp::min(mini_partition_duration_secs * 1_000_000, step as u64) as i64
-}
-
-/// Create a mini partition based on order
-fn create_mini_partition(
-    start_time: i64,
-    end_time: i64,
-    mini_partition_size: i64,
-    order_by: OrderBy,
-) -> Option<[i64; 2]> {
-    let duration = end_time - start_time;
-    if mini_partition_size >= duration {
-        return None;
-    }
-
-    let mini_partition = match order_by {
-        OrderBy::Desc => {
-            // For DESC order, mini partition at the end time
-            let mini_start = std::cmp::max(end_time - mini_partition_size, start_time);
-            [mini_start, end_time]
-        }
-        OrderBy::Asc => {
-            // For ASC order, mini partition at the start time
-            let mini_end = std::cmp::min(start_time + mini_partition_size, end_time);
-            [start_time, mini_end]
-        }
-    };
-
-    Some(mini_partition)
-}
-
-/// Generate regular partitions without histogram alignment
-fn generate_regular_partitions(
-    start_time: i64,
-    end_time: i64,
-    step: i64,
-    order_by: OrderBy,
-) -> Vec<[i64; 2]> {
-    let mut partitions = Vec::new();
-
-    match order_by {
-        OrderBy::Desc => {
-            let mut end = end_time;
-            while end > start_time {
-                let start = std::cmp::max(end - step, start_time);
-                partitions.push([start, end]);
-                end = start;
-            }
-        }
-        OrderBy::Asc => {
-            let mut start = start_time;
-            while start < end_time {
-                let end = std::cmp::min(start + step, end_time);
-                partitions.push([start, end]);
-                start = end;
-            }
-        }
-    }
-
-    partitions
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn print_partitions(title: &str, partitions: &[[i64; 2]]) {
-        println!("{title}");
-        for [start, end] in partitions.iter() {
-            let start = chrono::DateTime::from_timestamp_micros(*start)
-                .unwrap()
-                .with_timezone(&chrono::Local)
-                .format("%Y-%m-%d %H:%M:%S")
-                .to_string();
-            let end = chrono::DateTime::from_timestamp_micros(*end)
-                .unwrap()
-                .with_timezone(&chrono::Local)
-                .format("%Y-%m-%d %H:%M:%S")
-                .to_string();
-            println!("{start} - {end}");
-        }
-    }
-
-    // Actual start
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_no_mini_partition_where_step_is_equal_to_query_duration()
-     {
-        let start_time = 1753763400000000; // Tuesday, July 29, 2025 at 10:00:00 AM GMT+5:30
-        let end_time = 1753806600000000; // Tuesday, July 29, 2025 at 10:00:00 PM GMT+5:30
-        let min_step_seconds = 3600; // 60 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 30 minutes in microseconds
-
-        let step = 43200000000; // 12 hours in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS NO MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-07-29 10:00:00 - 2025-07-29 22:00:00
-        // HISTOGRAM PARTITIONS NO MINI PARTITION (DESC):
-        // 2025-07-29 21:30:00 - 2025-07-29 22:00:00
-        // 2025-07-29 10:00:00 - 2025-07-29 21:30:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions("HISTOGRAM PARTITIONS NO MINI PARTITION (ASC):", &partitions);
-
-        // Input
-        // 2025-07-29 10:00:00 - 2025-07-29 22:00:00
-        // HISTOGRAM PARTITIONS NO MINI PARTITION (ASC):
-        // 2025-07-29 10:00:00 - 2025-07-29 21:30:00
-        // 2025-07-29 21:30:00 - 2025-07-29 22:00:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_no_mini_partition_where_step_is_less_than_query_duration()
-     {
-        let start_time = 1753763400000000; // Tuesday, July 29, 2025 at 10:00:00 AM GMT+5:30
-        let end_time = 1753806600000000; // Tuesday, July 29, 2025 at 10:00:00 PM GMT+5:30
-        let min_step_seconds = 3600; // 60 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 30 minutes in microseconds
-
-        let step = 14400000000; // 4 hours in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS NO MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-07-29 10:00:00 - 2025-07-29 22:00:00
-        // HISTOGRAM PARTITIONS NO MINI PARTITION (DESC):
-        // 2025-07-29 21:30:00 - 2025-07-29 22:00:00
-        // 2025-07-29 17:30:00 - 2025-07-29 21:30:00
-        // 2025-07-29 13:30:00 - 2025-07-29 17:30:00
-        // 2025-07-29 10:00:00 - 2025-07-29 13:30:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions("HISTOGRAM PARTITIONS NO MINI PARTITION (ASC):", &partitions);
-
-        // Input
-        // 2025-07-29 10:00:00 - 2025-07-29 22:00:00
-        // HISTOGRAM PARTITIONS NO MINI PARTITION (ASC):
-        // 2025-07-29 10:00:00 - 2025-07-29 13:30:00
-        // 2025-07-29 13:30:00 - 2025-07-29 17:30:00
-        // 2025-07-29 17:30:00 - 2025-07-29 21:30:00
-        // 2025-07-29 21:30:00 - 2025-07-29 22:00:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_no_mini_partition_with_uneven_time_range()
-    {
-        let start_time = 1746074700000000; // Thursday, May 1, 2025 at 10:15:00 AM GMT+5:30
-        let end_time = 1746117300000000; // Thursday, May 1, 2025 at 10:05:00 PM GMT+5:30
-        let min_step_seconds = 3600; // 60 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 30 minutes in microseconds
-
-        let step = 14400000000; // 4 hours in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS NO MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-05-01 10:15:00 - 2025-05-01 22:05:00
-        // HISTOGRAM PARTITIONS NO MINI PARTITION (DESC):
-        // 2025-05-01 21:30:00 - 2025-05-01 22:05:00
-        // 2025-05-01 17:30:00 - 2025-05-01 21:30:00
-        // 2025-05-01 13:30:00 - 2025-05-01 17:30:00
-        // 2025-05-01 10:15:00 - 2025-05-01 13:30:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions("HISTOGRAM PARTITIONS NO MINI PARTITION (ASC):", &partitions);
-        // Input
-        // 2025-05-01 10:15:00 - 2025-05-01 22:05:00
-        // HISTOGRAM PARTITIONS NO MINI PARTITION (ASC):
-        // 2025-05-01 10:15:00 - 2025-05-01 13:30:00
-        // 2025-05-01 13:30:00 - 2025-05-01 17:30:00
-        // 2025-05-01 17:30:00 - 2025-05-01 21:30:00
-        // 2025-05-01 21:30:00 - 2025-05-01 22:05:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_no_mini_partition_with_uneven_time_range_2()
-     {
-        let start_time = 1746073920000000; // Thursday, May 1, 2025 at 10:02:00 AM GMT+5:30
-        let end_time = 1746074820000000; // Thursday, May 1, 2025 at 10:17:00 AM GMT+5:30
-        let min_step_seconds = 300; // 5 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 5 minutes in microseconds
-
-        let step = min_step; // 5 minutes in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS NO MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-05-01 10:02:00 - 2025-05-01 10:17:00
-        // HISTOGRAM PARTITIONS NO MINI PARTITION (DESC):
-        // 2025-05-01 10:15:00 - 2025-05-01 10:17:00
-        // 2025-05-01 10:10:00 - 2025-05-01 10:15:00
-        // 2025-05-01 10:05:00 - 2025-05-01 10:10:00
-        // 2025-05-01 10:02:00 - 2025-05-01 10:05:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions("HISTOGRAM PARTITIONS NO MINI PARTITION (ASC):", &partitions);
-
-        // Input
-        // 2025-05-01 10:02:00 - 2025-05-01 10:17:00
-        // HISTOGRAM PARTITIONS NO MINI PARTITION (ASC):
-        // 2025-05-01 10:02:00 - 2025-05-01 10:05:00
-        // 2025-05-01 10:05:00 - 2025-05-01 10:10:00
-        // 2025-05-01 10:10:00 - 2025-05-01 10:15:00
-        // 2025-05-01 10:15:00 - 2025-05-01 10:17:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_mini_partition_with_uneven_time_range_2() {
-        let start_time = 1746073920000000; // Thursday, May 1, 2025 at 10:02:00 AM GMT+5:30
-        let end_time = 1746074820000000; // Thursday, May 1, 2025 at 10:17:00 AM GMT+5:30
-        let min_step_seconds = 300; // 5 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 5 minutes in microseconds
-
-        let step = min_step; // 5 minutes in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-05-01 10:02:00 - 2025-05-01 10:17:00
-        // HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):
-        // 2025-05-01 10:16:00 - 2025-05-01 10:17:00
-        // 2025-05-01 10:15:00 - 2025-05-01 10:16:00
-        // 2025-05-01 10:10:00 - 2025-05-01 10:15:00
-        // 2025-05-01 10:05:00 - 2025-05-01 10:10:00
-        // 2025-05-01 10:02:00 - 2025-05-01 10:05:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-05-01 10:02:00 - 2025-05-01 10:17:00
-        // HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):
-        // 2025-05-01 10:02:00 - 2025-05-01 10:03:00
-        // 2025-05-01 10:03:00 - 2025-05-01 10:05:00
-        // 2025-05-01 10:05:00 - 2025-05-01 10:10:00
-        // 2025-05-01 10:10:00 - 2025-05-01 10:15:00
-        // 2025-05-01 10:15:00 - 2025-05-01 10:17:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_and_mini_partition_with_step_equal_to_query_duration()
-     {
-        let start_time = 1753763400000000; // Tuesday, July 29, 2025 at 10:00:00 AM GMT+5:30
-        let end_time = 1753806600000000; // Tuesday, July 29, 2025 at 10:00:00 PM GMT+5:30
-        let min_step_seconds = 1800; // 30 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 30 minutes in microseconds
-
-        let step = 43200000000; // 12 hours in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-07-29 10:00:00 - 2025-07-29 22:00:00
-        // HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):
-        // 2025-07-29 21:59:00 - 2025-07-29 22:00:00
-        // 2025-07-29 10:00:00 - 2025-07-29 21:59:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-07-29 10:00:00 - 2025-07-29 22:00:00
-        // HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):
-        // 2025-07-29 10:00:00 - 2025-07-29 10:01:00
-        // 2025-07-29 10:01:00 - 2025-07-29 22:00:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_and_mini_partition_with_step_less_than_query_duration()
-     {
-        let start_time = 1753763400000000; // Tuesday, July 29, 2025 at 10:00:00 AM GMT+5:30
-        let end_time = 1753806600000000; // Tuesday, July 29, 2025 at 10:00:00 PM GMT+5:30
-        let min_step_seconds = 3600; // 60 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 30 minutes in microseconds
-
-        let step = 14400000000; // 4 hours in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-07-29 10:00:00 - 2025-07-29 22:00:00
-        // HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):
-        // 2025-07-29 21:59:00 - 2025-07-29 22:00:00
-        // 2025-07-29 21:30:00 - 2025-07-29 21:59:00
-        // 2025-07-29 17:30:00 - 2025-07-29 21:30:00
-        // 2025-07-29 13:30:00 - 2025-07-29 17:30:00
-        // 2025-07-29 10:00:00 - 2025-07-29 13:30:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-07-29 10:00:00 - 2025-07-29 22:00:00
-        // HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):
-        // 2025-07-29 10:00:00 - 2025-07-29 10:01:00 - mini partition - Duplicates expected
-        // 2025-07-29 10:01:00 - 2025-07-29 13:30:00 - boundary - Duplicates expected
-        // 2025-07-29 13:30:00 - 2025-07-29 17:30:00 - 4hr -- Duplicates here
-        // 2025-07-29 17:30:00 - 2025-07-29 21:30:00 - 4hr -- Duplicates here
-        // 2025-07-29 21:30:00 - 2025-07-29 22:00:00 - boundary
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_and_mini_partition_with_step_greater_than_query_duration()
-     {
-        let start_time = 1746073920000000; // Thursday, May 1, 2025 at 10:02:00 AM GMT+5:30
-        let end_time = 1746161220000000; // Friday, May 2, 2025 at 10:17:00 AM GMT+5:30
-        let min_step_seconds = 3600; // 60 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 30 minutes in microseconds
-
-        let step = 86400000000; // 24 hours in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-05-01 10:02:00 - 2025-05-02 10:17:00
-        // HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):
-        // 2025-05-02 10:16:00 - 2025-05-02 10:17:00
-        // 2025-05-02 09:30:00 - 2025-05-02 10:16:00
-        // 2025-05-01 10:02:00 - 2025-05-02 09:30:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):",
-            &partitions,
-        );
-
-        // Input
-        // 2025-05-01 10:02:00 - 2025-05-02 10:17:00
-        // HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):
-        // 2025-05-01 10:02:00 - 2025-05-01 10:03:00
-        // 2025-05-01 10:03:00 - 2025-05-02 09:30:00
-        // 2025-05-02 09:30:00 - 2025-05-02 10:17:00
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-
-    #[test]
-    fn test_partition_generator_with_histogram_alignment_and_mini_partition_with_step_greater_than_query_duration_bug()
-     {
-        let start_time = 1752568628000000; // Thursday, May 1, 2025 at 10:02:00 AM GMT+5:30
-        let end_time = 1752741428000000; // Friday, May 2, 2025 at 10:17:00 AM GMT+5:30
-        let min_step_seconds = 18000; // 300 minutes in seconds
-        let min_step = min_step_seconds * 1_000_000; // 30 minutes in microseconds
-
-        let step = 72000000000; // 24 hours in microseconds
-
-        // Test Descending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            min_step,
-        );
-
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (DESC):",
-            &partitions,
-        );
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.last().unwrap()[0], start_time);
-        assert_eq!(partitions.first().unwrap()[1], end_time);
-
-        // Test Ascending
-        let partitions = generate_partitions_aligned_with_histogram_interval(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            min_step,
-        );
-
-        println!("partitions: {partitions:#?}");
-        print_partitions("Input", &[[start_time, end_time]]);
-        print_partitions(
-            "HISTOGRAM PARTITIONS WITH MINI PARTITION (ASC):",
-            &partitions,
-        );
-
-        // Verify full coverage of time range
-        assert_eq!(partitions.first().unwrap()[0], start_time);
-        assert_eq!(partitions.last().unwrap()[1], end_time);
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "enterprise")]
-mod enterprise_tests {
-    use super::*;
-
-    #[test]
-    fn test_partition_generator_with_streaming_aggregate_partition_window() {
-        let start_time: i64 = 1748513700000 * 1_000; // 10:15
-        let end_time: i64 = 1748528100000 * 1_000; // 14:15
-
-        let mini_partition_duration_secs = 60;
-        let step = 300000000; // 5 minutes
-
-        let partitions = generate_partitions_with_mini_partition(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            mini_partition_duration_secs,
-        );
-
-        let mut expected_partitions = vec![
-            [1748527200000000, 1748528100000000], // 14:00 - 14:15
-            [1748523600000000, 1748527200000000], // 13:00 - 14:00
-            [1748520000000000, 1748523600000000], // 12:00 - 13:00
-            [1748516400000000, 1748520000000000], // 11:00 - 12:00
-            [1748513700000000, 1748516400000000], // 10:15 - 11:00
-        ];
-        assert_eq!(partitions, expected_partitions);
-
-        let partitions_asc = generate_partitions_with_mini_partition(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Asc,
-            mini_partition_duration_secs,
-        );
-        expected_partitions.reverse();
-        assert_eq!(partitions_asc, expected_partitions);
-    }
-
-    #[test]
-    fn test_partition_generator_with_streaming_aggregate_aligned_boundary() {
-        // Test case where end_time is exactly aligned to window boundary
-        // This should not create empty partitions
-        let start_time: i64 = 1748966400000000; // Aligned to hour
-        let end_time: i64 = 1749153600000000; // Aligned to hour (52 hours later)
-
-        let mini_partition_duration_secs = 60;
-        let step = 300000000; // 5 minutes
-
-        let partitions = generate_partitions_with_mini_partition(
-            start_time,
-            end_time,
-            step,
-            OrderBy::Desc,
-            mini_partition_duration_secs,
-        );
-
-        // Verify no empty partitions exist
-        for [start, end] in &partitions {
-            assert!(
-                *end > *start,
-                "Partition [{start}, {end}] should not be empty"
-            );
-        }
-
-        // Verify full coverage
-        assert_eq!(
-            partitions.last().unwrap()[0],
-            start_time,
-            "First partition should start at start_time"
-        );
-        assert_eq!(
-            partitions.first().unwrap()[1],
-            end_time,
-            "Last partition should end at end_time"
-        );
-
-        // Verify partitions are contiguous
-        for i in 1..partitions.len() {
-            assert_eq!(
-                partitions[i][1],
-                partitions[i - 1][0],
-                "Partitions should be contiguous"
-            );
-        }
     }
 }
